Fix rendering of long items
More specifically, fix the subsequent items to not overlap the long
item's second (and additional) lines.

As part of this, change the highlight style from a different background
color to a rectangular border instead.  (This is to avoid having to
calculate the word wrapping as a separate step from the drawing, since
the background color has to get drawn before the text.)